+++ miniRT/inc/image.h
@@ -1,7 +1,7 @@
 #ifndef IMAGE_H
 # define IMAGE_H
 
-uint8_t		*get_pixel(mlx_image_t *img, int x, int y);
+uint32_t		*get_pixel(mlx_image_t *img, int x, int y);
 
 void		save_img(mlx_image_t *img, char *filename);
